Renamed SpawnServer and SpawnDaemon to ListenAndServe
